fix chat

+++ DigitalFUHubApi/Managers/ConnectionManager.cs
@@ -5,9 +5,9 @@ namespace DigitalFUHubApi.Managers
 	public class ConnectionManager : IConnectionManager	
 	{
 		//Save Hub type with connectionIds of a user
-		private static Dictionary<int, Dictionary<string, HashSet<string>>> data = new Dictionary<int, Dictionary<string, HashSet<string>>>();
+		private static Dictionary<long, Dictionary<string, HashSet<string>>> data = new Dictionary<long, Dictionary<string, HashSet<string>>>();
 
-		public void AddConnection(int userId,string hubName , string connectionId)
+		public void AddConnection(long userId,string hubName , string connectionId)
 		{
 			lock (data)
 			{
@@ -55,7 +55,7 @@ namespace DigitalFUHubApi.Managers
 			}
 		}
 
-		public HashSet<string>? GetConnections(int userId, string hubName)
+		public HashSet<string>? GetConnections(long userId, string hubName)
 		{
 			var connections = new HashSet<string>();
 			try
